feat: credible set quality control step
