+++ src/gentropy/method/susie_inf.py
@@ -6,10 +6,15 @@ from dataclasses import dataclass
 from typing import Any
 
 import numpy as np
+import pyspark.sql.functions as f
 import scipy.linalg
 import scipy.special
 from scipy.optimize import minimize, minimize_scalar
 from scipy.special import logsumexp
+
+from gentropy.dataset.ld_index import LDIndex
+from gentropy.dataset.study_index import StudyIndex
+from gentropy.dataset.study_locus import StudyLocus
 
 
 @dataclass
@@ -457,3 +462,47 @@ class SUSIE_inf:
                 )
             )
         return cred
+
+    @staticmethod
+    def credible_set_qc(
+        cred_sets: StudyLocus,
+        study_index: StudyIndex,
+        ld_index: LDIndex,
+        p_value_threshold: float = 1e-5,
+        purity_min_r2: float = 0.01,
+        ld_r2_threshold: float = 0.8,
+    ) -> StudyLocus:
+        """Filter credible sets by lead P-value and min-R2 purity, and performs LD clumping.
+
+        Args:
+            cred_sets (StudyLocus): StudyLocus object with credible sets to filter/clump
+            study_index (StudyIndex): StudyIndex object
+            ld_index (LDIndex): LDIndex object
+            p_value_threshold (float): p-value threshold for filtering credible sets, default is 1e-5
+            purity_min_r2 (float): min-R2 purity threshold for filtering credible sets, default is 0.01
+            ld_r2_threshold (float): LD R2 threshold for clumping, default is 0.8
+
+        Returns:
+            StudyLocus: Credible sets which pass filters and LD clumping.
+        """
+        df = (
+            cred_sets.df.withColumn(
+                "pValue", f.col("pValueMantissa") * f.pow(10, f.col("pValueExponent"))
+            )
+            .filter(f.col("pValue") <= p_value_threshold)
+            .filter(f.col("purityMinR2") >= purity_min_r2)
+            .drop("pValue")
+        )
+        cred_sets.df = df
+        cred_sets = (
+            cred_sets.annotate_ld(study_index, ld_index, r2_threshold=ld_r2_threshold)
+            .clump()
+            .filter(
+                ~f.array_contains(
+                    f.col("qualityControls"),
+                    "Explained by a more significant variant in high LD (clumped)",
+                )
+            )
+        )
+
+        return cred_sets
